Device restructure and addition of OpenVINO / Neural Compute Stick 2
Device restructure and addition of OpenVINO / Neural Compute Stick 2
