+++ README.md
@@ -99,8 +99,8 @@ Installation scripts and tutorials for setting up your Emergency Assistance Robo
 | ORDER | GUIDE | INFORMATION | AUTHOR | Status |
 | ----- | ----- | ----------- | ------ | ------ |
 | 1 | [Device 1](Devices/1/Documentation/Installation/Installation.md "Device 1") | Device 1 installation guide covering wheels installation. |  [Adam Milton-Barker](https://www.leukemiaresearchassociation.ai.com/team/adam-milton-barker "Adam Milton-Barker") | DEVELOPMENT | 
-| 2 | [Device 2](Devices/2/Documentation/Installation/Installation.md "Device 2") | Device 2 installation guide covering arm, Realsense and NCS2 installation |  [Adam Milton-Barker](https://www.leukemiaresearchassociation.ai.com/team/adam-milton-barker "Adam Milton-Barker") | DEVELOPMENT | 
-| 3 | [Device 3](Devices/3/Documentation/Installation/Installation.md "Device 3") | Device 3 installation guide covering camera pan/tilt  and NCS2 installation |  [Adam Milton-Barker](https://www.leukemiaresearchassociation.ai.com/team/adam-milton-barker "Adam Milton-Barker") | DEVELOPMENT | 
+| 2 | [Device 2](Devices/2/Documentation/Installation/Installation.md "Device 2") | Device 2 installation guide covering Realsense/OpenVINO and NCS2 installation |  [Adam Milton-Barker](https://www.leukemiaresearchassociation.ai.com/team/adam-milton-barker "Adam Milton-Barker") | DEVELOPMENT | 
+| 3 | [Device 3](Devices/3/Documentation/Installation/Installation.md "Device 3") | Device 3 installation guide covering arm, camera pan/tilt and NCS2 installation |  [Adam Milton-Barker](https://www.leukemiaresearchassociation.ai.com/team/adam-milton-barker "Adam Milton-Barker") | DEVELOPMENT | 
 
 &nbsp;
